feat: setting up models

+++ database/migrations/2024_10_22_204036_create_video_contents_table.php
@@ -14,7 +14,7 @@ return new class extends Migration
         Schema::create('video_contents', function (Blueprint $table) {
             $table->id();
             $table->string('type');
-            $table->unsignedTinyInteger('year');
+            $table->unsignedSmallInteger('year');
             $table->text('name');
             $table->text('description')->nullable();
             $table->timestamps();
